requirements.txt
asgiref==3.5.2
Django==4.0.4
django-ckeditor==6.4.2
django-js-asset==2.0.0
factory-boy==3.2.1
Faker==13.13.0
Pillow==9.1.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
whitenoise==6.1.0
